// src/enum/errors.enum.ts
export const enum Errors {
  asaNotFound = "Invalid asa id. Asa not found.",
  arcBadConfigured = "not configured correctly. Asset url is invalid.",
  arcBadConfiugredNoteIsUndefinedOrInvalid = 'not configured correctly. Field "note" is undefined or invalid.',
  invalidVersion = "Invalid version ~> 0",
  invalidMetadata = "Metadata is not a JSON",
}